Add a basic Kubernetes support
This commits is supposed to make it possible to use Kubernetes more or
less natively. There is third driver added to the selenoid that allows
creating pods and services for browsers that make it possible to use it
in a Kubernetes environment withoiht Dind.

Though, the support is very basic and there are a lot of thigs that
still have to be done:

- VNC support
- Resource allocation